change datapath handling. MUSIK_HOME_DIR define removed. a .Musik dir next to the exe is used if found now. fix same warnings

+++ trunk/wxMusik/src/Frames/Prefs/OptionFileAssocPanel.cpp
@@ -61,7 +61,7 @@ void OptionFileAssocPanel::DoFileAssociations()
 {
 
     bool bNotifyExplorer = false;
-    for(int i = 0; i < chklbFileAssocs->GetCount() ; i++)
+    for(size_t i = 0; i < chklbFileAssocs->GetCount() ; i++)
     {
         wxString sEntry =	chklbFileAssocs->GetString(i);
         wxArrayString arrElements;
